+++ src/components/staking/rewards/Activity.svelte
