+++ src/components/staking/rewards/Revenues.svelte
